spring-boot/src/main/java/com/zhangkai/spring/springboot/configbean/Book.java
package com.zhangkai.spring.springboot.configbean;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.stereotype.Component;

/**
 * 
 * @author and04
 *
 */
@Component
@ConfigurationProperties(prefix = "book")
public class Book {

	private String name;
	private String author;
	private String desc;

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public String getAuthor() {
		return author;
	}

	public void setAuthor(String author) {
		this.author = author;
	}

	public String getDesc() {
		return desc;
	}

	public void setDesc(String desc) {
		this.desc = desc;
	}

}
